+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/symbols/refs/JLocalVarReference.java
@@ -4,8 +4,6 @@
 
 package net.sourceforge.pmd.lang.java.symbols.refs;
 
-import java.lang.reflect.Modifier;
-
 import net.sourceforge.pmd.lang.java.ast.ASTVariableDeclaratorId;
 import net.sourceforge.pmd.lang.java.symbols.scopes.JScope;
 
@@ -19,7 +17,7 @@ import net.sourceforge.pmd.lang.java.symbols.scopes.JScope;
 public class JLocalVarReference extends AbstractCodeReference<ASTVariableDeclaratorId> implements JVarReference {
 
     // cannot be built from reflection, but a node is always available
-
+    private final boolean isFinal;
 
     /**
      * Constructor using the AST node.
@@ -28,12 +26,13 @@ public class JLocalVarReference extends AbstractCodeReference<ASTVariableDeclara
      * @param node           Node representing the id of the field, must be from an ASTLocalVariableDeclaration
      */
     public JLocalVarReference(JScope declaringScope, ASTVariableDeclaratorId node) {
-        super(declaringScope, node, node.isFinal() ? Modifier.FINAL : 0, node.getVariableName());
+        super(declaringScope, node, node.getVariableName());
+        this.isFinal = node.isFinal();
     }
 
 
     @Override
     public boolean isFinal() {
-        return Modifier.isFinal(modifiers);
+        return isFinal;
     }
 }
