+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/symbols/scopes/JScope.java
@@ -7,8 +7,12 @@ package net.sourceforge.pmd.lang.java.symbols.scopes;
 import java.util.Optional;
 import java.util.stream.Stream;
 
+import net.sourceforge.pmd.annotation.Experimental;
+import net.sourceforge.pmd.lang.java.symbols.refs.JCodeReference;
 import net.sourceforge.pmd.lang.java.symbols.refs.JMethodReference;
+import net.sourceforge.pmd.lang.java.symbols.refs.JSimpleTypeReference;
 import net.sourceforge.pmd.lang.java.symbols.refs.JSymbolicClassReference;
+import net.sourceforge.pmd.lang.java.symbols.refs.JTypeVariableReference;
 import net.sourceforge.pmd.lang.java.symbols.refs.JVarReference;
 import net.sourceforge.pmd.lang.java.symbols.scopes.internal.ImportOnDemandScope;
 import net.sourceforge.pmd.lang.java.symbols.scopes.internal.JavaLangScope;
@@ -17,37 +21,42 @@ import net.sourceforge.pmd.lang.java.symbols.scopes.internal.SingleImportScope;
 
 
 /**
- * Scope defined by the declaration of a type, a variable, a method.
- * Keeps track of the types, values, and methods accessible from their
- * simple name in the extent of the scope.
+ * A symbol table for a piece of Java code. Keeps track of the types, values, and
+ * methods accessible from their simple name in the extent of the scope.
  *
- * <p>Scopes are linked to their parent scope, and are only responsible for
- * the declarations occurring in their extent. When a scope can't find a
- * declaration, it delegates to its parent. This encodes shadowing and
- * hiding mechanisms transparently.
- *
- * <p>A class scope contains the field and method declarations of the class, and
- * is linked to an inherited scope that represents methods, types and values that
- * are inherited from its supertypes. It doesn't contain them directly.
- *
- *
- * <p>The following describes the most general form of the scope tree of a compilation unit.
- * Higher in the list means higher in the tree, i.e., lower precedence. Note that {@link ImportOnDemandScope}
- * has
- *
+ * <p>Each scope is linked to a parent scope, and keeps track of a particular set
+ * of declarations having the same relative precedence. When a scope is asked for
+ * the meaning of a name in a particular context (type name, method name, value name),
+ * it first determines if it has a declaration with a matching name.
  * <ul>
- * <li> {@link JavaLangScope}
- * <li> {@link ImportOnDemandScope}: never shadow anything, is shadowed by everything (see javadoc for why it's not the root)
- * <li> {@link SamePackageScope}: shadow imports-on-demands, is shadowed by single imports and lower
- * <li> {@link SingleImportScope}: shadows all of the above, is shadowed by type definitions of this compilation unit
- * </ul>*
+ *     <li>If there is one, it returns the {@link JCodeReference} representing the entity
+ *      the name stands for in the given context
+ *     <li>If there is none, it asks the same question to the parent scope recursively
+ *      and returns its result.
+ * </ul>
+ * This allows directly encoding shadowing and hiding mechanisms in the parent-child
+ * relationships.
  *
+ * <p>Each compilation unit defines a tree of scopes, somewhat corresponding to the
+ * logical structure of the compilation unit. The following describes the most general
+ * form of the upper part of the scope tree (before the first type declaration), in increasing
+ * order of precedence:
+ * <ul>
+ *     <li> {@link ImportOnDemandScope}: Types imported from a package or type by an import-on-demand,
+ *     and static method or field names imported from a type by a static-import-on-demand;
+ *     <li> {@link JavaLangScope}: Top-level types implicitly imported from {@literal java.lang};
+ *     <li> {@link SamePackageScope}: Top-level types from the same package, which are implicitly imported
+ *     <li> {@link SingleImportScope}: types imported by single-type-imports, and static methods and
+ *     fields imported by a single-static-import.
+ * </ul>
  *
- *
+ * <p>Note that the root of the tree is actually {@link JavaLangScope} for all compilation units, because
+ * we use a singleton to avoid duplicating it. The above precedence order is nevertheless respected.
  *
  * @author Clément Fournier
  * @since 7.0.0
  */
+@Experimental
 public interface JScope {
 
     /**
@@ -60,24 +69,28 @@ public interface JScope {
 
 
     /**
-     * Resolves the class referred to by this name. This must be a simple name.
+     * Resolves the type referred to by the given name. This must be a simple name,
+     * ie, parameterized types and array types are not available. Primitive types are
+     * also not considered because it's probably not useful.
+     *
+     * <p>The returned type reference may either be a {@link JSymbolicClassReference}
+     * or a {@link JTypeVariableReference}.
      *
      * @param simpleName Simple name of the type to look for
      *
-     * @return The class reference if it can be found, otherwise an empty optional
+     * @return The type reference if it can be found, otherwise an empty optional
      */
-    Optional<JSymbolicClassReference> resolveTypeName(String simpleName);
+    Optional<? extends JSimpleTypeReference<?>> resolveTypeName(String simpleName);
 
 
     /**
      * Finds all accessible methods that can be called with the given simple name
-     * on an implicit receiver in this scope. Several methods may be found, because
-     * they can have different arity and parameter types.
+     * on an implicit receiver in this scope. The returned methods may have different
+     * arity and parameter types.
      *
-     * @param simpleName Name
+     * @param simpleName Simple name of the method
      *
-     * @return An iterator enumerating methods with that name accessible
-     * from the implicit receiver in this scope.
+     * @return An iterator enumerating methods with that name accessible an applicable to the implicit receiver in this scope.
      */
     Stream<JMethodReference> resolveMethodName(String simpleName);
